requirements.txt
python-telegram-bot==20.7
web3==6.15.1
python-dotenv==1.0.1
aiohttp==3.9.3
aiosqlite==0.20.0
